Extend GoogleStorageResource to allow writing
In addition to being able to read objects in Google Cloud Storage
using Spring's Resource abstraction, we can now write resources
as well. GoogleStorageResource class now implements WritableResource
to support this. Note that if using @Value("gs://path-to/some-file.txt")
to inject a resource, it still needs to be cast from Resource to
WritableResource.

Fixes #72.