Polish credentials configuration
This commit renames AbstractCredentialsProperty to Credentials and offer
a way to initialize it with default scopes. Rather than creating a class
just for the purpose of setting them, all configuration properties
simply reuse the class now.

Also, this commit fixes a potential bug as we were previously relying on
the fact that the Credentials object was initialized when a property was
set. Arguably, we could initialize it by setting a scope (and not a
location, leading to a NPE).

Finally, this commit generates proper metadata for the credentials
sub-namespaces.