Stackdriver Logging Support
This initial implementation includes a Web MVC interceptor that
extracts the request trace ID from the X-Cloud-Trace-Context HTTP
header and makes it available to Stackdriver logging appenders.

This includes spring-cloud-gcp-logging and
spring-cloud-gcp-starter-logging modules with the corresponding unit
tests and readmes.

Reference documentation and a sample app are coming seprately.

Fixes #154.