Remove redundant starters
The "core" and "config" starters are really just passthrus for the
jars underneath them, and as such they are redundant. This change just
removes the empty layer.

Also remove mysql and postgres drivers from sql starter.
The user has to choose one or the other. The gcp starter just pulls
in the gcp features (which happen to include support for both
mysql and postgres).

User has to add JDBC driver and google socket factory. Seems like
a good compromise and simplifies the starters. Initializr can be
smart about creating new projects with the right socket factory
for the driver.