Improves lack of core credentials message
Currently, if no core credentials are set, but either the app doesn't
need them or they're set at the service level, an error message is
printed with a stack trace.

This leads users to believe there is an error with the app, when things
are actually fine. Only saying no core credentials are set is fine,
things will blow up when a service actually needs credentials but
doesn't have them.