Adds nativeHeaders to filtered headers
Using SI with Sleuth, we're finding that a nativeHeaders header with
a Map value is being added to the headers.
When a message is sent to Pub/Sub from a channel adapter, the map is
transformed to a string.
When the message returns from Pub/Sub, Sleuth casts the String to Map,
throwing an error in the process.

This is a stop-gap solution while we don't serialize Objects into
Strings, if we ever do.

Fixes #434