Fix logging autoconfig conditional
Currently, logging is automatically enabled for anyone running a
Spring Web MVC app, using any of our autoconfiguration that isn't
logging.

One remotely possible issue is when a user wants to use the Json
appender, doesn't use the logging starter but rather the autoconfigure
dependency with the google-cloud-logging-logback. In which case, they'll
have a ClassNotFoundException and should promptly add the missing
dependency.
We don't recommend using the logging bit without our starter, so this
is unlikely to begin with, and the solution is fairly straightforward,
so we don't think this is an issue.