Added new AckMode.AUTO_ACK
Added new AckMode.AUTO_ACK and implemented changes in
PubSubInboundChannelAdapter to process the new value.

AckMode.AUTO_ACK adds a new use case to PubSubInboundChannelAdapter in
which a message is only ack'ed if successfully processed. If an
exception occurs processing the message, the message will not be
nack'ed ... resulting in a redelivery delay by Google Pub / Sub.

Closes #850