Changed autoconfigured PubSubMessageConverter type
Changed autoconfigured PubSubMessageConverter type to
SimplePubSubMessageConverter to prevent a JacksonPubSubMessageConverter
from being used simply because Jackson was included in the project
dependencies.

There are various use cases where Jackson may be included, yet JSON is
not used. For example, serializing / deseralizing XML; XmlMapper is used
and is a subclass of ObjectMapper.