Move the gating flag into PubSubEmulator rule
The rule now emits a warning and performs no actions unless it.emulator
environment property is true. The implication is that you have to use
both it.pubsub and it.emulator flags together to get a meaningful test
run.

Also added a custom test rule in PubSubMessageChannelBinderTests to fail
with a meaningful message instead of the incomprehensible "Target" due
to binder getting a null constructor argument.