Use GcpPubSubHeaders.ORIGINAL_MESSAGE for manual acking.
GcpPubSubHeaders.ACKNOWLEDGEMENT is deprecated and unused, except in PubSubHeaderMapper.

I used BasicAcknowledgeablePubsubMessage and not ConvertedBasicAcknowledgeablePubsubMessage: since this header is only used for manual (n)acking, the exact implementation does not seem relevant. Besides, casting to PulledAcknowledgeablePubsubMessage is needed to get ackId, anyway. Might as well keep it as a clean base interface.