Fix GcpDatastoreEmulatorIntTests for projectId
The `projectId` is required for the `DatastoreOptions` and if we don't
specify it explicitly, the `ServiceOptions` super class falls back to
the `getDefaultProjectId()`, which eventually reads a projectId from the
`GOOGLE_APPLICATION_CREDENTIALS` files.

The `GcpDatastoreEmulatorIntegrationTests` fails on th CI server just
because there is no any credentials JSON file in that environment

* Specify `projectId` explicitly for the `DatastoreOptions.Builder`.
Use for this purpose a `GcpProjectIdProvider` bean from the context.