Added missing import to SamplerAutoConfiguration
without this change the sampler configuration is not read. By default in Sleuth core, sampling is not necessary cause we're not sending spans anywhere. If you're using Zipkin or any other tracing system to which you want to send spans, you should consider adding sampling.
with this change we're explicitly (which is the proper thing to do) adding import to the SamplerAutoConfiguration so that the sampling logic is setup. It happens BEFORE the TraceAutoConfiguration, so the Sampler.NEVER_SAMPLE will not be registered.

fixes spring-cloud/spring-cloud-sleuth#1409