Use Ilford Spring Cloud BOM (#3)
Upgrade to Ilford.

Since Spring Cloud GCP is no longer included in the Spring Cloud BOM for Ilford, we can actually use the Spring BOM ourselves in our parent POM.

The version of Sleuth is being overwritten to the old one because the new one causes our integration tests to fail.